tests/i: tidy test utilities away out of sight
