+++ tests/integration/utils/test_flow_writer.py
@@ -18,20 +18,16 @@
 https://github.com/cylc/cylc-flow/pull/2740#discussion_r206086008
 
 And yes, these are unit-tests inside a functional test framework thinggy.
+
 """
-from pathlib import Path
+
 from textwrap import dedent
 
-import pytest
-
-from . import (
+from .flow_writer import (
     _write_header,
     _write_setting,
     _write_section,
-    suiterc,
-    _rm_if_empty,
-    _poll_file,
-    _expanduser
+    suiterc
 )
 
 
@@ -118,41 +114,3 @@ def test_suiterc():
             [[bar]]
                 pub = beer
     ''').strip() + '\n'
-
-
-def test_rm_if_empty(tmp_path):
-    """It should remove dirs if empty and suppress exceptions otherwise."""
-    path1 = Path(tmp_path, 'foo')
-    path2 = Path(path1, 'bar')
-    path2.mkdir(parents=True)
-    _rm_if_empty(path1)
-    assert path2.exists()
-    _rm_if_empty(path2)
-    assert not path2.exists()
-    _rm_if_empty(path1)
-    assert not path1.exists()
-
-
-@pytest.mark.asyncio
-async def test_poll_file(tmp_path):
-    """It should return if the condition is met."""
-    path = tmp_path / 'file'
-    await _poll_file(path, exists=False)
-    path.touch()
-    await _poll_file(path, exists=True)
-
-
-def test_expanduser():
-    """It should expand ~ and $HOME."""
-    assert _expanduser('a/~/b') == Path('a/~/b').expanduser()
-    assert _expanduser('a/$HOME/b') == Path('a/~/b').expanduser()
-    assert _expanduser('a/${HOME}/b') == Path('a/~/b').expanduser()
-
-
-def test_flow(run_dir, flow, one_conf):
-    """It should create a flow in the run directory."""
-    reg = flow(one_conf)
-    assert Path(run_dir / reg).exists()
-    assert Path(run_dir / reg / 'suite.rc').exists()
-    with open(Path(run_dir / reg / 'suite.rc'), 'r') as suiterc:
-        assert 'scheduling' in suiterc.read()
